Rename path of FollowButton
