+++ front/src/components/UserProCard.vue
@@ -2,7 +2,7 @@
 import ProfilePicture from "./ProfilePicture.vue";
 import ProfileStat from "./ScudoTheming/ProfileStat.vue";
 import Alert from "./ScudoTheming/Alert.vue";
-import FollowButton from "./FollowButton.vue";
+import FollowButton from "./ScudoTheming/FollowButton.vue";
 import Text from "./ScudoTheming/Text.vue";
 
 const props = defineProps({
